More <double> to <float>

+++ qtsynthitem.cpp
@@ -82,7 +82,7 @@ void QtSynthItem::disconnectAll()
 }
 
 // Set the data column, dataset minumum and dataset maximum values
-void QtSynthItem::setData(std::vector<double> *data, std::vector<double> *mins, std::vector<double> *maxes)
+void QtSynthItem::setData(std::vector<float> *data, std::vector<float> *mins, std::vector<float> *maxes)
 {
     DatasetCommand cmd;
     cmd.data = data;
