+++ qtsynthitem.h
@@ -35,9 +35,9 @@ public:
     Q_INVOKABLE virtual void disconnect(QtSynthItem *parent);
     Q_INVOKABLE virtual void disconnectAll();
 
-    virtual void setData(std::vector<double>* data,
-                             std::vector<double>* mins,
-                             std::vector<double>* maxes);
+    virtual void setData(std::vector<float>* data,
+                             std::vector<float>* mins,
+                             std::vector<float>* maxes);
     virtual Frame process();        // called every audio sample
     virtual void step();            // called every new data value (step)
     virtual void controlProcess();  // called every process block
